requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.4
Flask==1.1.2
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
line-bot-sdk==1.18.0
MarkupSafe==1.1.1
pytz==2020.5
requests==2.25.1
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
Werkzeug==1.0.1
